PR 69: Updated calls to PHP_curl to go through a new curlHelper class so we can
Updated calls to PHP_curl to go through a new curlHelper class so we can control, from a single place, which parameters should be set by default. Set the CURLOPT_SSL_VERIFYPEER=true and CURLOPT_SSL_VERIFYHOST=2
Added comments to the commented out code which was using file_get_contents() to migrate to curlHelper if this is ever revived.

Related work items: #145