Merged PR 477: UI - training_page_polling
UI - training_page_polling. Training page polls every 2 seconds even when it doesn't need to.

If we haven't uploaded anything for training, or if the training has completed, we stop polling. When we upload a new file the poll star again.
in other case poll is each 2 seconds

Related work items: #827