Merged PR 554: Scaling2: ignore a server until its first heartbeat succeeds
As soon as a server comes online with lots of free capacity it would get allocated all the new chats.
We need to ignore it until the first time the heartbeat call returns a 200 so that if it is misconfigured we don't return a bunch of chat error 500s.

Related work items: #944, #945