Merged PR 610: Scaling 2: AIML servers only serve AIIDs that they report on registration
When an AIML server registers it provides a list of AIs.
The API/Controller will serve AIML chat for that AI to that server only.
If no AIML server has registered that AIID then an exception is thrown.
N.B. there are a bunch of automatic method re-arranges in ServerMetadata.java!

Related work items: #1165