Merged PR 627: Fix bug 1088 in scaling - No requests are being routed to AIML
Linking the chit-chat bot failed because a class selector was finding the bots that had a specific class to mark them as selected bots. This was then passed as a list of which bots should be linked. This class was deleted and partially replaced with an attribute so the selector failed.

The selector now searches for the attribute.

Cherry picked from Pull request 626.

Related work items: #1088