Merged PR 713: Added support for chat bot affinity.
Added support for chat bot affinity.
A conversation will now get "locked" to the bot that initially had the highest score, until the score of the responses for the bot fall below the threshold. This is true even if other bots have higher scores when the conversation is locked. As soon as the locked bot's score falls below the threshold, we then choose the highest confidence bot response and lock to that one.
Added topic to the chat state.

Related work items: #1346, #1340, #1341