Merged PR 758: Add explicit row-locking to SPs used for AI status updates
Having a REPEATABLE_READ transaction is not sufficient for data integrity.
When reading data inside a transaction, we have to add "FOR UPDATE" to the sql so that MySQL knows to lock these rows until they are updated and the transaction is committed.
This change fixes a bug with updates from different back-end servers that occur at the same time - both status-update and db-sync calls.

Related work items: #1426