Merged PR 755: Prevents an AI from being deleted in the backends if there is already a published but associated with i
Prevents an AI from being deleted in the backends if there is already a published but associated with it.
For now ay published but will remain in the store (and purchaseable/usable) until we model the lifecycle, but for now this prevents anyone using the bot from getting an error.

Related work items: #1423