Merged PR 773: UI - limit number of bots that can be linked to 5.
UI - limit number of bots that can be linked to 5.
Change made to AI creation and Settings flows (moved code to single function).
Now during the creation of an AI, when you "purchase" a bot it will no be automatically added to the skills, so the limit can then be validated when it's added.

Related work items: #1460