Merged PR 782: Fixed the training flow after an intent has been updated.
Fixed the training flow after an intent has been updated.
We now handle more training states and instruct the API to re-upload the training file to the backends (which will contain the new intent data).
Not the prettiest code, but in the "style" of the existing one...

Related work items: #1242