Merged PR 796: Set score=1 if we're sucessfuly processing an intent.
Set score=1 if we're sucessfuly processing an intent.
This applies for intent-multiturn: when we're in the middle of prompting the user about the intent's entities, we now set the score to 1.0 (previously was always 0.0).
Note that the the intent is *recognized* initially by WNET, we still send the confidence level that WNET returns.

Related work items: #1502