Merged PR 803: Moving the UI logging to elasticsearch.
Moving the UI logging to elasticsearch.
This is just a stopgap solution because 1. we don't log that much from the UI; 2. We need to use a logreader (running outside of PHP) to prevent the UI from breaking if ES is not available.
Both of these problems exist in the old platform, just switcing logging endpoint.

Related work items: #1523