Merged PR 874: Fix for closing the responses (connections) to the server.
Fix for closing the responses (connections) to the server.
When there is a connection issue (or a response that is not HTTP_OK), the response was not being fully consumes, which would leave the connections open. Since JerseyClient's connection pooling defaults to 2 max connections per host, it's likely that this was causing issues trying to upload logs if there was a real connectivity problem, so it would not be able to swiftly recover.
Also updated JerseyClient's connection pool configuration to more realistic values (this affects all modules that use it).

Related work items: #1722