Merged PR 893: Scaling3: timeslicing of training on multiple servers
There is now an AI status table that doubles up as a command queue.
There are two DB scripts: one is an upgrade script from master to scaling3.
The other is the DB creation script that generates a clean DB. (+users +default_data)
API training commands now get intercepted and the DELETE and TRAIN tasks get queued.
One QueueProcessor runs for each of WNET, RNN. This class checks the database for queued tasks and sends the commands to the backend, updating the queue. Failed commands get re-queued.
Commands go to the connected, training-capable server with the lowest loading factor.
If a back-end server calls back to say that a bot has gone from TRAINING state to QUEUED then we put this bot back into the training queue. This enables interleaved timeslicing for training on the RNN. (changes required on the back-end)

N.B. A number of non-essential bits are missing in this change but commented in the code as TODOs. e.g. requeuing policy, timeout policy.

Related work items: #1492, #1491, #1495, #1271, #1270, #1493