Merged PR 900: Prevent the UI from sending logs to ES until we have a good strategy
Prevent the UI from sending logs to ES until we have a good strategy that doesn't block the interface in case of failure. For now just writing it to the error log which will end up in the Apache log. We should later use a logreader to push that into ES (and when we have meaningful logging on the UI as well).

Related work items: #1288