Merged PR 905: Scaling: fix queue processor to use Timer schedule rather than Thread.sleep
(This is ground-work for logging server connection status to the database.)
The queue processor now runs on a scheduled timer rather than variable length Thread.sleep commands.
Inside the timer is logic to determine whether to run the queue code or wait for the next tick so that we have variable length timings anyway.
We process the queue more often when things are happening, less often when they are not.

Related work items: #1678