Merged PR 912: Added a healthcheck endpoint with a ping operation to respond 200 OK
Added a healthcheck endpoint with a ping operation to respond 200 OK when queries.
This is to support NLBs querying the state of the service. We can't use another endpoint due to 1)most require authentication, and unauthorised will be seen by the NLB as service down; 2) the only unauthenticated calls are the /ui and these will incur db access which is not needed for a simple ping