Merged PR 924: Added password/salt and dev/client tokens to the Hutoma account
Added password/salt and dev/client tokens to the Hutoma account. Also made it active.
The default password is 'password' for the account, which will be changed on each environment to something different.
This will allow logging in with the hutoma account so we can publish bots originating from Hutoma.

Related work items: #1599