Merged PR 947: Scaling3: requeue policy takes error-code into account
Merge davidMG/scaling3_requeue_policy to scaling3

For a delete command, if the back-end sends back a 404 then there is no point trying again.
In the case of training, if the back-end sends a 400 or a 404 then we set the AI status to AI_ERROR and don't try again.
For all other errors we requeue the command to try again later.

Related work items: #1760