Merged PR 967: Updated the logging of the heartbeat status field to always be a String
Updated the logging of the heartbeat status field to always be a String to make sure the index will contain consistent data. If the first document added to the index has this field set to integer, then it sets the type, affecting any other logs being added later where this field is set to String.

Cherry picked from Pull request 966.

This is required for PROD otherwise after the index is set then there's no way of changing the type

Related work items: #1819