Merged PR 1037: Scaling3.2: don't do any training recovery for the first few minutes after the API boots
For the first n minutes (n=2) after the queue processor starts up training recovery is skipped.
This is to ensure that if the API goes down but the back-end servers do not it does not enthusiastically re-allocate all the training jobs to different servers. It waits a couple of minutes for existing back-end servers to reclaim their own training jobs.

Related work items: #1850