Merged PR 1042: Scaling3.2: dont reset status in dbsync when an AI was training
If we have a status of "training" in the database then we keep "training" if the server says "queued" or "ready to train".
This is because one server may think that the AI is "queued" or "ready to train" while _another_ server is training the AI.
Also, if the back-end server reboots then it will report an AI that was previously training as "queued". If we updated our status to "queued" then we would never properly queue the AI for training.

Related work items: #1496