Merged PR 1052: Scaling 3.2: hide requeuing of bot training from the UI (and any API caller)
When a bot is queued for training we still report it as queued.
If it gets requeued, i.e. for any reason the server trains for a while and then pauses, we still report to outside world that the bot is "training".
Effectively we are hiding most of the internal scheduling from the user.

Related work items: #1500