Merged PR 1078: Extracted the bot details page to be able to be called externally (and
Extracted the bot details page to be able to be called externally (and unauthenticated).
Basically this page can be called at <site>/console/botcardDetail.php?botId=<botId>
and will render the page with the details. When called from the botstore (within the console) the behaviour should remain the same, and it displays within an iframe.
Fixed some bugs I found and did the least amount of changes I could to get this working as we'll need to do "some" refactoring.
There are still some issues that were there originaly (like trying to purchase within the botstore opens the botstore in the frame, resulting in 2 menus, etc) that I've left for later.

Related work items: #1970