Merged PR 1097: API will now abort an intent if entity not provided.
API will now abort an intent if entity not provided.
Before the API would wait for an entity to appear in the question, even after
it stopped prompting for an entity.
But that leads to issues like bug #1982.

The behaviour is changed so that if an entity is not provided after the
prompts, the intent is cleared.

Related work items: #1982