Merged PR 1095: Updated the CreateAI flow to show just the purchased botskills instead of a
Updated the CreateAI flow to show just the purchased botskills instead of a view of the botstore.

I experimented with displaying the details card as an overlaid modal, but the way the code is structured makes it pretty difficult without having to refactor large sections of code, so I've opted for the smallest and simplest solution.

Note that during the creation flow if you open a details card and go back you will still be able to create the AI with the data you've entered previously, but if you start navigating around the botstore you will lose the ability to go back (unless you press back enough times to get there - which isn't a good UX). So we'll need to have a further change to create the bot when the user naivgates away of the creation flow

Related work items: #1994