Merged PR 1104: Session-based Persistent Entity Values
This is a V0 feature for internal use only, required for the Nokia work and demo this week. Persistence is defined as chat length for now to make it simple/binary but when we further design we'll have some sort of timeout. Entities are set as persistent via a flag in the database.

When a 'persistent' entity is populated it's value is stored for that user and session in the DB, and populated whenever we trigger an intent that uses the entity. eg.

> Tell me the Description for alarm 7424
> "blah..."
> Tell me the Meaning
> "blah..."

So that we don't have to specify the alarm again.

- Adding database storage for persisting entity values for a chat session.
- Adding stored procedures for reading/writing values.
- Update entities table to allow flagging an entity as persistent.

- Updating tests to verify non-persistent variables still behave as normal.
- Adding tests to verify that an intent can be fulfilled by a previous set persistent value.

Related work items: #1962, #1963, #1961