Merged PR 1091: Added support for listing bots from all the categories with a single API call.
Added support for listing bots from all the categories with a single API call.
The API returns a dictionary keyed by the category name, and containing a list of the bots per category.
To understand if there are more items on each carousel, one just needs to query with max=NUM+1. This means if you want to show NUM=10 items and want to display a button to see more, you need to request 11 items. For every category with 11 items the button should be displayed. I didn't implement the paging strategy on this request as this is a different scenario from paging on a single category.

Related work items: #1974