Merged PR 1117: Fixed "See more" functionality to open the carousel on the same frame.
Fixed "See more" functionality to open the carousel on the same frame.
The menu category switching is done through an event to support both situations where the menu can be there or not (in fact it will also support adding another menu as long as it uses the same id).
Updated the existing paint eventto use plain JS's events and not jquery

Related work items: #2014