Merged PR 1138: Don't allow an in-use entity to be deleted
An entity is in use if any intent refers to it.

UI: This will check for the 409 error from API and redirect to error page with code 327. Some refactoring of the common error handling to use a switch statement.

API: First lookup to see if the entity named exists and belongs to the dev. If not then return 404 (new StoredProcedure getEntityIdForDev). Then if so, do a DELETE on the entity if there is no entry for that entity in the intent_variables table (update SP deleteEntity).

DB: Update to stored procedures, allow entityUser to SELECT from intent_variables.

Related work items: #1475