Merged PR 1196: Support intent-only AIs.
Support intent-only AIs.
In the UI side, it now correctly calls updateTraining to trigger the API to upload the training materials (intents) to the backend.
On the API side it allows training to be updated from AI_UNDEFINED as long as there's no user training file (in that case then the state could never be AI_UNDEFINED unless there was an error uploading the user training file). Updated case where w were returning 500 if the user would call updateTraining on an AI in the wrong state - changed it to Bad Request

Related work items: #2130