Merged PR 1212: Changed chat logic so that MinP (confidence threshold) is now stored internally
Changed chat logic so that MinP (confidence threshold) is now stored internally as part of the chat state and is not passed around between UI and API. This simplifies user API calls as well.
For the first chat interaction (no state yet) the confidence threshold is read from the AI's parameters defined by the user. After taht we re-use the value in the state.

Related work items: #2150