Merged PR 1216: Added Hotjar support to the console.
Added Hotjar support to the console.
All pages now include the hotjar script as long as the environment variable with the tracking ID is present.
Updated the Intercom script to also only be present if the environment variable is set (and to read it from the env instead of having it hardcoded - allowing us to have a separate id for testing).

Note that the hotjar tracking also needed to be added to the pages that are served within an iframe (card details and botstore list).

Related work items: #2181