Merged PR 1228: Ignore a status update from the back-end that tells us a bot is now STOPPED
In the Scaling3 world, the 'stopped' status means user-requested stop ...
but the back-end servers will return 'stopped' if they stop training due to a 409 error so we need to _not_ update the status in this case.
For other cases when the user tells the API to tell the back-end to stop then we update the 'stopped' status on the API side.

Related work items: #2140