Merged PR 1213: Added alter db script to include all the db alterations required to be done on
Added alter db script to include all the db alterations required to be done on the database during the next deployment to PROD. Since we always cut a release branch, this will only contain the changes from the previous deployment. This makes it much easier to deploy to PROD (even automate it) as we won't need to be doing a diff between create scripts and then manually create the alter script.

Related work items: #2216