Merged PR 1247: Changed chat logic to take into account the individual minP value for each AI.
Changed chat logic to take into account the individual minP value for each AI.
For now (to maintain the existing behaviour) we will still take as the result the AI that provided the highest score, but this makes it simpler to later have some sort of weighted scores if we want.
Note that this applies only to WNET as with RNN we're always hard-coding the score, and with AIML we wither return a high score or 0.

Related work items: #2244