Merged PR 1244: API endpoint for Hutoma Facebook app incoming webhook calls
This is an endpoint for the Facebook platform to communicate with the API via webhooks.
1. the endpoint must be able to verify to Facebook that it is able to process calls (token + challenge)
2. it will accept POST notifications whenever a user sends a facebook message to a subscribed page.
For now, this endpoint will verify and accept messages but will not do anything with the messages apart from log them. As part of later tasks the API will respond to chat messages.

N.B. this can only go into master after the beta release.

Related work items: #2119