Merged PR 1292: Facebook: check granted permissions with every config call
When users connect to Facebook they get to review the list of requested permissions and _edit_ the grant list.
They can also re-edit the list at any time on web Facebook.
As a result we need to check the grant list against our required list and feed anything missing back to the user.

Also in this change: the permission request list comes from the API and is no longer hard-coded in the console.

Related work items: #2280