Merged PR 1299: Facebook: stealthed the "integrations" page in preparation for Beta3
Remove link from sidebar.menu.js to integrations.
To go to the Facebook integrations page you can edit the URL directly to
/console/integrations.php

##Why?
We are "stealth" launching the feature, meaning that all the code goes live but the user doesn't get to see the link.
This is so that we can submit our Facebook integration app for review without exposing our users to all sorts of errors that result from using an unverified, unpublished Facebook App.
Once the review is complete we can test the feature and "unstealth" it for instant joy. The "unstealth" code change will be small and low risk.

Related work items: #2339