Merged PR 1302: Change to support multiple variables of the same entity type during chat.
Change to support multiple variables of the same entity type during chat.
We now store the entity that is being requested on a multi-turn conversation, by setting a flag in MemoryVariable. This will then get picked up in the next turn, so we know which variable was - as now it's not enough just to know the name of the variable, but you need to know the instance itself.
Had to refactor the intent handling code as this was getting too messy after all the previous iterations when we added more features.
Added a UT (big as a house) to validate the evaluation of the entities to prompt.

Related work items: #2337