Merged PR 1315: Fixed webhook url regex validation. It was expecting the 3rd part of an IP
Fixed webhook url regex validation. It was expecting the 3rd part of an IP address to be alpha only, which is obviously incorrect. Also fixed the URI validation tahthad the same problem. This problem has been in the code for a while (exists in PROD) but I guess we haven't had validated IP addresses with that code yet

Related work items: #2391