Merged PR 1331: Fix for some symbols being rendered incorrectly.
Fix for some symbols being rendered incorrectly.
The cleanChat() function was performing replacements which would mess up the string to display, added a new stripHtml() method to use the browser's engine to obtain the text from within an element in memory.
TODO: need to check if there are any security implications if injecting a script, although the element being only created in memory.