Merged PR 1335: Added chatlogs download page and associated API endpoint.
Added chatlogs download page and associated API endpoint.
The API queries the ES log server for chatlogs for a particular AI, and presents that as a download to the user, using the requested format. Currently the API only presents the option to download as .csv, although the API already suports JSON which we can decide to expose later.
The UI uses bootstrap's date-picker (included) to allow the user to easily select the time window. Currently there is no limit on number of days or file size, we may want to enforce a 30 day span later to make the downloads quicker.

Related work items: #2417, #2416