Merged PR 1300: Facebook: unstealth integrations config page
### N.B. This change should only go into master after B3, to then be cherry-picked into PROD after the Facebook team favourably reviews our integration app.

Re-added the link to integrations in the sidebar menu
Changed the name of the sidebar menu javascript file to avoid semi-permanent caching that will hide this feature from our users.

### Why?
We are “stealth” launching the feature, meaning that all the code goes live but the user doesn’t get to see the link. This is so that we can submit our Facebook integration app for review without exposing our users to all sorts of errors that result from using an unverified, unpublished Facebook App.
Once the review is complete we can test the feature and “unstealth” it for instant joy. This is the "unstealth" change.

Related work items: #2339