Merged PR 1353: Added initial view of charts on the insights page (chat sessions and
Added initial view of charts on the insights page (chat sessions and interactions).
We should decide at some stage how is the best option to provide the user the option to define the time interval for the charts - list of common intervals (last 7 days, last 30 days, etc), free-form, etc, and how this is laid out on the page. For now, for simplicity, it's locked down to displaying the last 30 days.

![image.png](https://hutoma.visualstudio.com/_apis/git/repositories/304a971b-36f8-4319-b2ab-e5666b2b37de/pullRequests/1353/attachments/image%20(3).png)

Also note that the chart will auto-scale based on the number of data points (so the bars will become slimmer when where are datapoints for more days)

Related work items: #2453