Merged PR 1364: Deliver rich-content from webhook to Facebook
We have no way of returning rich content from the back-end servers yet but we can now pass rich-content from a webhook to Facebook.

A webhook response in this format:
~~~~
{
  "text":"This is sent to a normal text client.",
  "facebook":{
    "type":"image",
    "payload":{
      "url":"https://facebook.users.get.a.picture.instead/img/shirt.png"
    }
  }
}
~~~~
will deliver normal text to an API or console caller but will display an image on Facebook.

So far we only support audio, video, file and image attachments. Generic templates (i.e. lists with buttons and things) will be part of a later task.

Related work items: #2425