Merged PR 1371: Create separate hmac_secret for webhooks, only sign HTTPS webhooks
Only sign HTTPS webhooks as HTTP is inherently insecure.

New secret value is stored in the AI table as hmac_secret, defaults to NULL.
This is auto-generated the first time we try and use it if it is NULL.
I ruled out adding it to createAi flow as this is already an 11 argument call and SP which makes updating error-prone.

Created a new random hex generator function in tools, based on SecureRandom call.

Exposing this to API call and UI work to come later.
Updated unit-tests for Webhook.
Updated DB scripts for new SPs and database structure.

Related work items: #2256, #2255