Merged PR 1379: Fixed the chart date interval filtering.
Fixed the chart date interval filtering.
ES's post filtering was not filtering out the the dates correctly when using aggregations, so now the date filter is part of the 'must' clause.
Changed charts to display the whole time frame, and not to auto-resize based on the datapoints, to make the charts more consistent.
Fixed issue when there're no data points to display -> prints message instead of trying to render the chart (which would fail).
Changed position of the Insights entry on the menu

![image.png](https://hutoma.visualstudio.com/_apis/git/repositories/304a971b-36f8-4319-b2ab-e5666b2b37de/pullRequests/1379/attachments/image.png)

Related work items: #2456, #2461