Merged PR 1382: Enforce password restrictions when creating a new account.
Enforce password restrictions when creating a new account.
On the JS, if the restrictions are not met we always display it as a 'weak' password.
The PHP then re-validates the restrictions before attempting to call the API.

Obviously this isn't the best UX as the user should not have to press Submit if any of the entries is invalid, but this is just temporary until we sort out this whole page

Related work items: #1381, #2470