Merged PR 1393: Communicate errors  better when Facebook fails to connect
Keep the result of a Facebook connect action in the PHP session so that we can display it to the user on the next page load.
The API now returns error 409 as a result of FB connect when things go wrong or there is a warning to display.

Related work items: #2482