Merged PR 1406: Fixed the month being offset by one.
Fixed the month being offset by one.
This was caused by JS in its infinite wisdom attributing January=0, which basically makes any conversion from ISO8601 not direct as months are in the [1,12] range.
Set the bar width to a fixed value as with group % it looks weird if you have only a few data points (which will happen to all devs at the beginning)

Related work items: #2494