Merged PR 1415: Fixed incorrect encoding of comma for prompts.
Fixed incorrect encoding of comma for prompts.
It didn't support adding a prompt with a comma anyway, so used a more standard encode/decodeURIcomponent to build a comma separated string from an array. As far as I can tell this never worked before, we just didn't test with commas.

![image.png](https://hutoma.visualstudio.com/_apis/git/repositories/304a971b-36f8-4319-b2ab-e5666b2b37de/pullRequests/1415/attachments/image.png)

Related work items: #2517