Merged PR 1405: Added bot auto-purchase during account creation.
Added bot auto-purchase during account creation.
This is controlled by the API_AUTO_PURCHASE_BOTLIST env var taking a comma separated list of Bot IDs (not AIIDs, as the botId is the id for a published bot).
Fixed an issue with the account creation due to the removal of the user's client token

Related work items: #2511