Merged PR 1418: Fix page clash when the user connects more than one bot to a page
When a user tries to connect a bot to a page that is already connected to another bot then we switch off the old integration to enable the new one.

This fixes a bug where about 50% of the time chat responses where going to the old bot and failing anyway because of Auth errors.

Related work items: #2484