Merged PR 1433: Added complexity validation to password reset plus a few UI fixes.
Added complexity validation to password reset plus a few UI fixes.
Text color for login/registration/reset pages is now #A0A0A0 instead of the darkish green gray that was more difficult to see on a black background (especially for errors).
Updated cookie policy link on registration page.
Updated errors to show the error as H3 and the text as normal text (to make it consistent with how it was already in some cases).