Merged PR 1447: Facebook integration: support for postback buttons
The user can now send a button: type=postback to us in a webhook and we will relay it to Facebook.
We accept postback messages from the Facebook webhook and send the payload text to the chat engine as a query. Effectively, clicking on a postback button means sending us a line of text that is defined in the bot-developer's webhook. This _can_ be used to trigger an intent, which could easily return more postback buttons etc.

Related work items: #2571