Merged PR 1450: Facebook integration: expand intent-variable prompt to use button template
If the chat-response is a variable prompt
and it is not for a system entity
and the entity has between 1 and 12 values then
Send a button template to Facebook that allows the user to click on the required entity value.
Button templates can only have 1-3 buttons so we send multiple messages to Facebook with different sets of buttons.
i.e. this change adds support for multiple-message responses to Facebook.

Related work items: #2584