Merged PR 1473: Added console support for user-defined default chat responses.
Added console support for user-defined default chat responses.
Note that we're only supporting one entry for this on the Settings page, which means that although the rest of the system supports multiple entries, we'll be limiting this for now to a single entry for users.
Once we refactor the UI code to make it simpler to handle multiple entries we can then restore full support for this functionality.

Related work items: #2659