Merged PR 1500: Facebook quick-replies from a webhook
Support for accepting quick-replies from webhooks and sending them to Facebook.
This introduces a new webhook JSON format where the facebook node is no longer an `attachment` but a `message` that _contains_ an `attachment`.
Both formats are supported but only the newer format can be used to send quick-replies.

Related work items: #2716