Merged PR 1502: API documentation: Turn on enunciate as part of API build
Maven update: The enunciate pages are generated and are part of the WAR file we deploy. This is now a part of `mvn package` and not `mvn site`.

As a result it will run on CI builds, daily builds and be part of generated Docker images automatically.

We have to tell the Enunciate/Swagger module where to find our end points in enunciate config.

Updated enunciate to latest version (2.9.1).

I played with JavaDoc before this and found some errors, so fixed these up too (as found by mvn javadoc:javadoc).

Related work items: #2741