Merged PR 1493: Changed botstore (carousel) to use Mustache templates.
Changed botstore (carousel) to use Mustache templates.
Renamed .mst to .mustache files so IntelliJ can read them as templates.
On the botstore carousel item, I've changed the "see more" button to always show regardless if there are more items or not (and renamed to "see all") as the existing logic for displaying the button was completely broken and just "working" by a strike of luck... If we decide that we really want to hide this button we should write that functionality from scratch