Merged PR 1526: Facebook customisations: Greeting and Get Started
Adds an editable 'greeting' free text field that is shown to the user when first interacting with a bot on Facebook.
Also adds an editable payload that is sent to the bot when the user clicks Get Started on first interacting with a bot on Facebook.
These are set in the integration page of the console UI. The data is stored in the metadata field in the integration record.
Most of the code changes are to account for the fact that the metadata doesn't get completely blanked out every time the bot-owner disconnects and reconnects a bot from Facebook.

Related work items: #2711, #2710