Merged PR 1561: Fix null check on both PHP and JS for bot icon.
Fix null check on both PHP and JS for bot icon.
On PHP, this was causing the bot creation flow to break when an existing purchased bot had no icon associated.
On JS, it failed presenting the bot card with on the same condition.

Note that this code wasn't changed during the ui refactoring, we're now just passing json structures around as sent by the API without having intermediate structures, and this is likely to start revealing some similar issues.

Related work items: #2848