Merged PR 1604: Fix for not being able to delete an entity.
Fix for not being able to delete an entity.
Since the delete code relies on the original index of the array, and we display it sorted, th order was changed, so we needed to keep the index before sorting.
Also, there was a problem with obtaining the name of the entity to delete to present on the modal (this is not refactoring, as it shows as well in PROD) which was fixed as well.

Related work items: #2892