Merged PR 1607: Force entity labels to be unique per intent.
Force entity labels to be unique per intent.
We were allowing similar labels if the entity type was different, but this causes a problem because we generate a map that is keyed on the entity label.
For instance if you had sys.places with label 'A' and sys.names labeled 'A' as well, when we generate the map these are going to clash. So we need to make labels unique per intent.
Updated the check on the UI, and added validation on the API side. Also added a WARN for already saved labels since we'll need to update these (manually?)

Also removed another file that survived the merge

Related work items: #2891