Merged PR 1627: Fix for password not resetting after the ui refactoring.
Fix for password not resetting after the ui refactoring.
The problem was that previously we were depending on data to be present in the session, whereas now this is not the case. So we need to extract the userId from the password reset token, then ue that for requesting the API for the password to be reset.
Added some more logging for failure conditions around that area.

Related work items: #2942