Merged PR 1634: Fixes for publishing errors.
Fixes for publishing errors.
The flow is confusing as we call checkInput (through an event handler) AND displayAlertMessage to validate the fields. The latter was not passing the id of the element failing validation, causing a JS error. There was also function name clashing with a different part of publishing (developer) so renamed the method to make sure we're calling the right one.
We should really consolidate the two methods but since this is a release blocker I'm not confident that bigger changes wouldn't cause further regressions, so keeping it just as a spot fix.

Related work items: #2945