Merged PR 1641: Revert "Re-adding the check to prevent publishing when there are linked bots."
Revert "Re-adding the check to prevent publishing when there are linked bots."

Reverting this to maintain current PROD behaviour. Regardless on how we want to handle this situation, we can now do it exclusively from the front end side since we now send the data required to make that decision.

Related work items: #2688