Merged PR 1670: Fix memory leak caused by tracking threads inside singleton classes
ThreadSubPool was designed to be able to take _n_ threads from the main thread pool, use them together and then close them together. The problem is that we are using ThreadSubPool in singleton classes where we create new threads periodically without every closing the sub-pool. The threading bit works fine but we end up with an ever-growing list of threads that we are tracking and this keeps some resources alive too long.
This change splits ThreadSubPool into two: the new ThreadSubPool does no tracking and can be used everywhere (including in singletons). TrackedThreadSubPool tracks threads and provides auto-close functionality just like the old ThreadSubPool did. This should only be used when tracking is really necessary, and never in singletons or objects with singleton lifespans.
Also in this change:
QueueProcessor is a singleton class and makes use of AiQueueServices, which in turn makes use of TrackedThreadSubPool, so it has to use a Provider<> to create AiQueueServices on demand and keep their lifespan down to a single operation.

Before:
![Before the fix](http://www.dixiak.com/hal/memleak.png)

After:
![After the fix](http://www.dixiak.com/hal/afterfix.png)

Related work items: #2854, #2909